requirements.txt
alembic==0.7.5.post2
aniso8601==0.92
BeautifulSoup==3.2.1
beautifulsoup4==4.3.1
blinker==1.2
cov-core==1.15.0
coverage==3.7.1
cssmin==0.2.0
decorator==3.4.0
enum34==1.0.4
execnet==1.3.0
factory-boy==2.5.2
Flask==0.10.1
Flask-Admin==1.0.7
Flask-Assets==0.10
Flask-Mail==0.8.2
Flask-Migrate==1.3.1
Flask-RESTful==0.3.2
Flask-Script==2.0.5
Flask-SQLAlchemy==2.0
Flask-WTF==0.9.0
flup==1.0.2
gunicorn==0.17.2
html5lib==0.99
itsdangerous==0.24
Jinja2==2.7.3
lxml==2.2.4
Mako==1.0.1
MarkupSafe==0.23
pathlib==1.0.1
pbr==0.10.8
psycopg2==2.5
py==1.4.26
PyExecJS==1.1.0
pyScss==1.3.4
pytest==2.7.0
pytest-cov==1.8.1
pytest-xdist==1.11
pytz==2013b0
requests==1.2.3
six==1.4.1
speaklater==1.3
SQLAlchemy==1.0.0
sqlalchemy-migrate==0.9.6
sqlparse==0.1.15
Tempita==0.5.1
webassets==0.10.1
Werkzeug==0.10.4
WTForms==1.0.3
